Refactor mock filesystem to dedicated testing package
This commit reorganizes the mock filesystem implementation for better
structure and clarity:

• Move MockFileSystem from internal/common to internal/common/testing
• Update package documentation to clearly indicate testing purpose
• Update all imports across security and verification modules to use
  qualified import alias commontesting
• Remove build constraint //go:build test from types_test.go as it's
  no longer needed
• Maintain all existing functionality while improving code organization
