+++ internal/runner/security/security_path_test.go
@@ -5,7 +5,7 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/isseis/go-safe-cmd-runner/internal/common"
+	commontesting "github.com/isseis/go-safe-cmd-runner/internal/common/testing"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -15,14 +15,14 @@ func TestValidator_ValidateDirectoryPermissions_CompletePath(t *testing.T) {
 
 	tests := []struct {
 		name        string
-		setupFunc   func(*common.MockFileSystem)
+		setupFunc   func(*commontesting.MockFileSystem)
 		dirPath     string
 		shouldFail  bool
 		expectedErr error
 	}{
 		{
 			name: "valid directory with secure path",
-			setupFunc: func(fs *common.MockFileSystem) {
+			setupFunc: func(fs *commontesting.MockFileSystem) {
 				// Create secure directory hierarchy
 				fs.AddDir("/usr", 0o755)
 				fs.AddDir("/usr/local", 0o755)
@@ -35,7 +35,7 @@ func TestValidator_ValidateDirectoryPermissions_CompletePath(t *testing.T) {
 		},
 		{
 			name: "directory with world-writable intermediate directory",
-			setupFunc: func(fs *common.MockFileSystem) {
+			setupFunc: func(fs *commontesting.MockFileSystem) {
 				fs.AddDir("/usr", 0o755)
 				fs.AddDir("/usr/local", 0o777) // World writable - insecure!
 				fs.AddDir("/usr/local/etc", 0o755)
@@ -48,7 +48,7 @@ func TestValidator_ValidateDirectoryPermissions_CompletePath(t *testing.T) {
 		},
 		{
 			name: "directory with group-writable intermediate directory owned by non-root",
-			setupFunc: func(fs *common.MockFileSystem) {
+			setupFunc: func(fs *commontesting.MockFileSystem) {
 				fs.AddDir("/opt", 0o755)
 				fs.AddDirWithOwner("/opt/myapp", 0o775, 1000, 1000) // Group writable, owned by non-root
 				fs.AddDir("/opt/myapp/etc", 0o755)
@@ -61,7 +61,7 @@ func TestValidator_ValidateDirectoryPermissions_CompletePath(t *testing.T) {
 		},
 		{
 			name: "directory with root group write owned by root",
-			setupFunc: func(fs *common.MockFileSystem) {
+			setupFunc: func(fs *commontesting.MockFileSystem) {
 				fs.AddDirWithOwner("/usr", 0o775, 0, 0) // Root group writable, owned by root - allowed
 				fs.AddDir("/usr/local", 0o755)
 				fs.AddDir("/usr/local/etc", 0o755)
@@ -73,7 +73,7 @@ func TestValidator_ValidateDirectoryPermissions_CompletePath(t *testing.T) {
 		},
 		{
 			name: "directory with non-root group write owned by root",
-			setupFunc: func(fs *common.MockFileSystem) {
+			setupFunc: func(fs *commontesting.MockFileSystem) {
 				fs.AddDirWithOwner("/usr", 0o775, 0, 1) // Non-root group writable, owned by root - prohibited
 				fs.AddDir("/usr/local", 0o755)
 				fs.AddDir("/usr/local/etc", 0o755)
@@ -86,7 +86,7 @@ func TestValidator_ValidateDirectoryPermissions_CompletePath(t *testing.T) {
 		},
 		{
 			name: "directory owned by current user",
-			setupFunc: func(fs *common.MockFileSystem) {
+			setupFunc: func(fs *commontesting.MockFileSystem) {
 				fs.AddDir("/home", 0o755)
 				fs.AddDirWithOwner("/home/user", 0o755, uint32(os.Getuid()), uint32(os.Getgid())) // Owned by current user
 				fs.AddDir("/home/user/config", 0o755)
@@ -96,7 +96,7 @@ func TestValidator_ValidateDirectoryPermissions_CompletePath(t *testing.T) {
 		},
 		{
 			name: "directory owned by different non-root user",
-			setupFunc: func(fs *common.MockFileSystem) {
+			setupFunc: func(fs *commontesting.MockFileSystem) {
 				fs.AddDir("/home", 0o755)
 				fs.AddDirWithOwner("/home/user", 0o755, 2000, 2000) // Owned by different non-root user (UID 2000)
 				fs.AddDir("/home/user/config", 0o755)
@@ -119,7 +119,7 @@ func TestValidator_ValidateDirectoryPermissions_CompletePath(t *testing.T) {
 		},
 		{
 			name: "root directory with insecure permissions",
-			setupFunc: func(fs *common.MockFileSystem) {
+			setupFunc: func(fs *commontesting.MockFileSystem) {
 				// Replace default secure root with insecure one
 				fs.RemoveAll("/")
 				fs.AddDirWithOwner("/", 0o777, 0, 0) // World-writable root - insecure!
@@ -135,7 +135,7 @@ func TestValidator_ValidateDirectoryPermissions_CompletePath(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			// Create fresh mock filesystem for each test
-			testMockFS := common.NewMockFileSystem()
+			testMockFS := commontesting.NewMockFileSystem()
 			testValidator, err := NewValidatorWithFS(config, testMockFS)
 			require.NoError(t, err)
 
@@ -163,14 +163,14 @@ func TestValidator_ValidateCompletePath_SymlinkProtection(t *testing.T) {
 
 	tests := []struct {
 		name        string
-		setupFunc   func(*common.MockFileSystem)
+		setupFunc   func(*commontesting.MockFileSystem)
 		path        string
 		shouldFail  bool
 		expectedErr error
 	}{
 		{
 			name: "path with symlink component should be rejected",
-			setupFunc: func(fs *common.MockFileSystem) {
+			setupFunc: func(fs *commontesting.MockFileSystem) {
 				// Create secure directory hierarchy, but skip /usr/local as we'll replace it with a symlink
 				fs.AddDir("/usr", 0o755)
 
@@ -188,7 +188,7 @@ func TestValidator_ValidateCompletePath_SymlinkProtection(t *testing.T) {
 		},
 		{
 			name: "path with symlink target directory should be rejected",
-			setupFunc: func(fs *common.MockFileSystem) {
+			setupFunc: func(fs *commontesting.MockFileSystem) {
 				// Create secure directory hierarchy
 				fs.AddDir("/usr", 0o755)
 				fs.AddDir("/usr/local", 0o755)
@@ -208,7 +208,7 @@ func TestValidator_ValidateCompletePath_SymlinkProtection(t *testing.T) {
 		},
 		{
 			name: "secure path with no symlinks should pass",
-			setupFunc: func(fs *common.MockFileSystem) {
+			setupFunc: func(fs *commontesting.MockFileSystem) {
 				// Create completely normal directory hierarchy
 				fs.AddDir("/usr", 0o755)
 				fs.AddDir("/usr/local", 0o755)
@@ -221,7 +221,7 @@ func TestValidator_ValidateCompletePath_SymlinkProtection(t *testing.T) {
 		},
 		{
 			name: "AddSymlink should fail when path already exists",
-			setupFunc: func(fs *common.MockFileSystem) {
+			setupFunc: func(fs *commontesting.MockFileSystem) {
 				// Create directory first
 				fs.AddDir("/usr", 0o755)
 				fs.AddDir("/usr/existing", 0o755)
@@ -239,7 +239,7 @@ func TestValidator_ValidateCompletePath_SymlinkProtection(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			// Create fresh mock filesystem for each test
-			testMockFS := common.NewMockFileSystem()
+			testMockFS := commontesting.NewMockFileSystem()
 			testValidator, err := NewValidatorWithFS(config, testMockFS)
 			require.NoError(t, err)
 
@@ -270,7 +270,7 @@ func TestValidator_ValidatePathComponents_EdgeCases(t *testing.T) {
 
 	tests := []struct {
 		name        string
-		setupFunc   func(*common.MockFileSystem)
+		setupFunc   func(*commontesting.MockFileSystem)
 		path        string
 		shouldFail  bool
 		expectedErr string
@@ -283,7 +283,7 @@ func TestValidator_ValidatePathComponents_EdgeCases(t *testing.T) {
 		},
 		{
 			name: "single level directory",
-			setupFunc: func(fs *common.MockFileSystem) {
+			setupFunc: func(fs *commontesting.MockFileSystem) {
 				fs.AddDir("/test", 0o755)
 			},
 			path:       "/test",
@@ -291,7 +291,7 @@ func TestValidator_ValidatePathComponents_EdgeCases(t *testing.T) {
 		},
 		{
 			name: "path with double slashes",
-			setupFunc: func(fs *common.MockFileSystem) {
+			setupFunc: func(fs *commontesting.MockFileSystem) {
 				fs.AddDir("/usr", 0o755)
 				fs.AddDir("/usr/local", 0o755)
 			},
@@ -300,7 +300,7 @@ func TestValidator_ValidatePathComponents_EdgeCases(t *testing.T) {
 		},
 		{
 			name: "empty path components handled",
-			setupFunc: func(fs *common.MockFileSystem) {
+			setupFunc: func(fs *commontesting.MockFileSystem) {
 				fs.AddDir("/usr", 0o755)
 				fs.AddDir("/usr/local", 0o755)
 			},
@@ -312,7 +312,7 @@ func TestValidator_ValidatePathComponents_EdgeCases(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			// Create fresh mock filesystem for each test
-			testMockFS := common.NewMockFileSystem()
+			testMockFS := commontesting.NewMockFileSystem()
 			testValidator, err := NewValidatorWithFS(config, testMockFS)
 			require.NoError(t, err)
