+++ internal/runner/security/types_test.go
@@ -1,5 +1,3 @@
-//go:build test
-
 package security
 
 import (
