+++ internal/verification/errors_test.go
@@ -4,7 +4,7 @@ import (
 	"errors"
 	"testing"
 
-	"github.com/isseis/go-safe-cmd-runner/internal/common"
+	commontesting "github.com/isseis/go-safe-cmd-runner/internal/common/testing"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -161,7 +161,7 @@ func TestValidateProductionConstraints(t *testing.T) {
 func TestSecurityConstraintsInManager(t *testing.T) {
 	t.Run("production mode with strict security enforces constraints", func(t *testing.T) {
 		_, err := newManagerInternal("/custom/dir",
-			withFSInternal(common.NewMockFileSystem()),
+			withFSInternal(commontesting.NewMockFileSystem()),
 			withFileValidatorDisabledInternal(),
 			withCreationMode(CreationModeProduction),
 			withSecurityLevel(SecurityLevelStrict),
@@ -173,7 +173,7 @@ func TestSecurityConstraintsInManager(t *testing.T) {
 	})
 
 	t.Run("testing mode with relaxed security allows custom directories", func(t *testing.T) {
-		mockFS := common.NewMockFileSystem()
+		mockFS := commontesting.NewMockFileSystem()
 		mockFS.AddDir("/custom", 0o755)
 		mockFS.AddDir("/custom/dir", 0o755)
